fix: move did to didcomm dir and update module to PublicModule
