+++ apps/vs-agent/src/controllers/public/didcomm/DidWebController.ts
