+++ apps/vs-agent/src/controllers/public/index.ts
@@ -1,3 +1,3 @@
-export * from './did'
+export * from './didcomm'
 export * from './invitation'
 export * from './self-vtr'
